ko_build.Read: Check if underlying directory has been deleted.
This has proven to be really difficult to test, since we need to mutate
the state in a way that produces an initial valid config, but then
mutates the underlying source to be invalid.

For now, trusting that the error is being properly returned back up.
This was somewhat spot checked using https://go.dev/play/p/yIIp6_Y__OL
to make sure we can error.Is correctly.

Signed-off-by: Billy Lynch <billy@chainguard.dev>